+++ packages/plugin-contain/README.md
@@ -1,0 +1,3 @@
+# @jimp/plugin-contain
+
+contain an image.
